feature: pool and analogs calculate

+++ app/database/tables/adjustment.py
@@ -1,6 +1,6 @@
 import uuid
 
-from sqlalchemy import Column, ForeignKey, Integer, Numeric
+from sqlalchemy import Column, Float, ForeignKey, Numeric
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import relationship
 
@@ -13,19 +13,19 @@ class Adjustment(Base):
     guid = Column(UUID(as_uuid=True), default=uuid.uuid4, primary_key=True, index=True, unique=True)
     apartment_guid = Column(UUID(as_uuid=True), ForeignKey("apartment.guid"))
     apartment = relationship("Apartment", back_populates="adjustment", uselist=False)
-    trade = Column(Integer, nullable=False, default=-450)
+    trade = Column(Float, nullable=False, default=-450)
     price_trade = Column(Numeric(7, 2), nullable=False)
-    floor = Column(Integer, nullable=False)
+    floor = Column(Float, nullable=False)
     price_floor = Column(Numeric(7, 2), nullable=False)
-    apt_area = Column(Integer, nullable=False)
+    apt_area = Column(Float, nullable=False)
     price_area = Column(Numeric(7, 2), nullable=False)
-    kitchen_area = Column(Integer, nullable=False)
+    kitchen_area = Column(Float, nullable=False)
     price_kitchen = Column(Numeric(7, 2), nullable=False)
-    has_balcony = Column(Integer, nullable=False)
+    has_balcony = Column(Float, nullable=False)
     price_balcony = Column(Numeric(7, 2), nullable=False)
-    distance_to_metro = Column(Integer, nullable=False)
+    distance_to_metro = Column(Float, nullable=False)
     price_metro = Column(Numeric(7, 2), nullable=False)
-    quality = Column(Integer, nullable=False)
+    quality = Column(Float, nullable=False)
     price_final = Column(Numeric(7, 2), nullable=False)
 
     analog_calculated_guid = Column(UUID(as_uuid=True), ForeignKey("sub_query.guid"))
